pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "asyncio>=4.0.0",
    "beautifulsoup4>=4.13.5",
    "google-generativeai>=0.8.5",
    "lxml>=6.0.1",
    "nodriver>=0.47.0",
]
